refactor: shared bind-mount-provisioning module + unified spawn primitive
Two depth refactors against the code that shipped earlier in this PR.

1. Hoist `BindMount`, `resolveParentGitMount`, and `defaultImageName` out
   of `sandboxes/docker.ts` into `internal/bind-mount-provisioning.ts`.
   These describe what host paths a bind-mount sandbox provider needs
   to mount and what its default image tag should be — they're not
   docker-specific. The (in-flight) Podman bind-mount provider will be
   the second adapter consuming this module; without the seam it would
   either duplicate the mount-resolution logic or reach across into a
   sibling provider.

   Tests: filesystem-fixture coverage for `resolveParentGitMount` (regular
   checkout vs pointer file vs malformed pointer) co-located with the
   already-shipped `defaultImageName` cases. The old
   `sandboxes/docker.test.ts` is removed — there's nothing docker-specific
   left in it.

2. Unify the two near-identical sub-process primitives. `internal/git-spawn`
   and `sandboxes/docker.runHostProcess` were both `spawn → collect
   stdout/stderr → decode → return`, differing only in trivia (cwd
   required vs not, throw-on-non-zero vs return-exit-code, line streaming).
   Both will need the same future cross-cutting work — abort signals (#12)
   and idle timeouts (#11) — so concentrating the spawn semantics in one
   place pays off forward.

   `internal/host-process.spawnHost` is the new primitive: never throws
   on non-zero exit (caller decides), supports an `onLine` streaming
   option, takes optional `cwd`. `runGit` becomes a thin throw-on-fail
   wrapper. Docker's container ops call `spawnHost` directly.

3. Document the env-resolver's provider-blind assumption. The
   `HOST_ONLY_ENV_KEYS` blocklist is correct for bind-mount and
   isolated sandbox providers but wrong for the no-sandbox provider
   (#18), where the agent runs on the host and HOME/PATH/USER from
   `process.env` are exactly what's wanted. Comment notes that env
   policy should move to a per-provider declaration when #18 lands;
   no behavior change today (one adapter — bind-mount via docker —
   currently exercises the filter).

Validated end-to-end against real Docker: both deterministic shell-agent
demos (head + merge-to-head) pass with no functional change relative to
pre-refactor.
